backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "structure-ocr-backend"
version = "0.1.0"
description = "FastAPI backend for the Structure-OCR document extraction app"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.8.0",
  "pymupdf>=1.24.0",
  "pillow>=10.4.0",
]

[project.optional-dependencies]
ocr = [
  "paddleocr>=3.4.0",
  "paddlex[serving]>=3.4.1",
]
